+++ src/demo/kotlin/com/loadingbyte/cinecred/demo/VirtualDesktop.kt
@@ -604,7 +604,7 @@ class SpreadsheetEditorVirtualWindow(private val file: Path, skipRows: Int) : Fa
 
     override val title = "${file.name} \u2013 " + l10nDemo("screencast.spreadsheetEditor.title")
 
-    val matrix = CsvFormat.read(file).first.single().drop(skipRows).map { it.cells.toMutableList() }
+    val matrix = CsvFormat.read(file, "").first.single().drop(skipRows).map { it.cells.toMutableList() }
     var colWidths = intArrayOf()
 
     private var selectedCell: Point? = null